requirements.txt
streamlit
pymongo
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
pydantic
requests
groq
tiktoken
python-multipart
sentence-transformers
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
